fix delete day nanti belom bisa refersh

+++ resources/views/user.blade.php
@@ -92,7 +92,7 @@
                             </td>
                             <td class="px-6 py-4 flex items-center gap-2">
                                 <button type="button" class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800" onclick="showEditUserModal(${user.id})">Edit</button>
-                                <button type="button" class="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900">Delete</button>
+                                <button type="button" class="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900" onclick="deleteUser(${user.id})">Delete</button>
 
                             </td>
                         </tr>
@@ -205,33 +205,43 @@
         })
     }
 
-    function deleteUser() {
-        $.ajax({
-            url: "{{ url('/deleteUser') }}",
-            type: 'DELETE',
-            data: {
-                id: $('#userId').val(),
-                _token: "{{ csrf_token() }}",
-            },
-            success: function(response) {
-                if (response.success == true) {
-                    hideFlowBytesModal('edit-user-modal');
-                    getAllUsers();
-                    swal.fire({
-                        title: 'Success',
-                        text: response.message,
-                        icon: 'success',
-                        confirmButtonText: 'OK'
-                    })
-                } else {
-                    swal.fire({
-                        title: 'Error',
-                        text: response.message,
-                        icon: 'error',
-                        confirmButtonText: 'OK'
-                    })
-                }
-
+    function deleteUser(id) {
+        swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    url: "{{ url('/deleteUser') }}",
+                    type: 'DELETE',
+                    data: {
+                        id: id,
+                        _token: "{{ csrf_token() }}",
+                    },
+                    success: function(response) {
+                        if (response.success == true) {
+                            getAllUsers();
+                            swal.fire({
+                                title: 'Success',
+                                text: response.message,
+                                icon: 'success',
+                                confirmButtonText: 'OK'
+                            })
+                        } else {
+                            swal.fire({
+                                title: 'Error',
+                                text: response.message,
+                                icon: 'error',
+                                confirmButtonText: 'OK'
+                            })
+                        }
+                    }
+                })
             }
         })
     }
